:sparkles: Ticket Service - 공통모듈로 Dto 이동
