+++ common/src/main/java/com/ipho/common/dto/ConfirmSeatEvent.java
@@ -1,4 +1,4 @@
-package com.ipho.ticketservice.application.event.dto;
+package com.ipho.common.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -9,10 +9,9 @@ import java.util.UUID;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class CancelTicketEvent {
+public class ConfirmSeatEvent {
 
     private UUID eventId;
     private String seatNumber;
     private Double price;
-
 }
