+++ common/src/main/java/com/ipho/common/dto/SeatBookingEvent.java
@@ -1,19 +1,17 @@
-package com.ipho.ticketservice.application.event.dto;
+package com.ipho.common.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class TicketMakingEvent {
+public class SeatBookingEvent {
     private UUID ticketId;
-    private UUID seatId;
-    private String eventName;
+    private UUID eventId;
+    private Long userId;
     private String seatNumber;
-    private BigDecimal price;
 }
