+++ common/src/main/java/com/ipho/common/dto/TicketMakingEvent.java
@@ -1,17 +1,19 @@
-package com.ipho.ticketservice.application.event.dto;
+package com.ipho.common.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class SeatBookingEvent {
+public class TicketMakingEvent {
     private UUID ticketId;
-    private UUID eventId;
-    private Long userId;
+    private UUID seatId;
+    private String eventName;
     private String seatNumber;
+    private BigDecimal price;
 }
