+++ ticket-service/src/main/resources/application.yml
@@ -21,7 +21,7 @@ spring:
       group-id: ${spring.application.name}-group
       auto-offset-reset: earliest
       key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
-      value-deserializer: org.springframework.kafka.support.serializer.JsonDeserializer
+      value-deserializer: org.springframework.kafka.support.serializer.ErrorHandlingDeserializer
       properties:
         spring.json.trusted.packages: '*'
         specific-value-deserializer: org.springframework.kafka.support.serializer.JsonDeserializer
