+++ ticket-service/src/test/java/com/ipho/ticketservice/service/TicketServiceMockTest.java
@@ -4,9 +4,9 @@ import com.ipho.ticketservice.application.dto.TicketInfoDto;
 import com.ipho.ticketservice.infrastructure.client.SeatClientService;
 import com.ipho.ticketservice.presentation.request.TicketRequestDto;
 import com.ipho.ticketservice.presentation.response.TicketResponseDto;
-import com.ipho.ticketservice.application.event.dto.CancelTicketEvent;
-import com.ipho.ticketservice.application.event.dto.SeatBookingEvent;
-import com.ipho.ticketservice.application.event.service.EventProducer;
+import com.ipho.common.dto.CancelTicketEvent;
+import com.ipho.common.dto.SeatBookingEvent;
+import com.ipho.ticketservice.application.event.EventProducer;
 import com.ipho.ticketservice.application.service.TicketService;
 import com.ipho.ticketservice.domain.model.Ticket;
 import com.ipho.ticketservice.domain.repository.TicketRepository;
