+++ ticket-service/src/test/java/com/ipho/ticketservice/service/TicketServiceTest.java
@@ -1,11 +1,11 @@
 package com.ipho.ticketservice.service;
 
 import com.ipho.ticketservice.application.dto.TicketInfoDto;
-import com.ipho.ticketservice.application.event.dto.CancelTicketEvent;
-import com.ipho.ticketservice.application.event.dto.SeatBookingEvent;
-import com.ipho.ticketservice.application.event.dto.TicketMakingEvent;
-import com.ipho.ticketservice.application.event.dto.TicketTopic;
-import com.ipho.ticketservice.application.event.service.EventProducer;
+import com.ipho.common.dto.CancelTicketEvent;
+import com.ipho.common.dto.SeatBookingEvent;
+import com.ipho.common.dto.TicketMakingEvent;
+import com.ipho.ticketservice.application.event.TicketTopic;
+import com.ipho.ticketservice.application.event.EventProducer;
 import com.ipho.ticketservice.application.service.TicketService;
 import com.ipho.ticketservice.domain.model.Ticket;
 import com.ipho.ticketservice.domain.model.TicketStatus;
